Forward compatibility with littleEndianWords in Blob API
We use Big Endian for Spectra and Little Endian for the rest.
To be consistent we'll need to fix one of them. We'll probably
change everything to Big Endian, but that would break this SDK
right away. So for forward compatibility the API will temporarily
handle littleEndianWords flag when we start switching.